+++ apps/web/src/pages/TalentRequests/components/TalentRequestReferralDialog/TalentRequestCreateTrackedUserDialog.tsx
@@ -8,13 +8,12 @@ import {
   getFragment,
   graphql,
 } from "@gc-digital-talent/graphql";
-import { commonMessages, formMessages } from "@gc-digital-talent/i18n";
-import { Button, Dialog, Link } from "@gc-digital-talent/ui";
+import { Button, Dialog } from "@gc-digital-talent/ui";
 import { toast } from "@gc-digital-talent/toast";
 
-import useRoutes from "~/hooks/useRoutes";
 import { getFullNameLabel } from "~/utils/nameUtils";
 
+import ReferralDialogFooter from "./ReferralDialogFooter";
 import ReferralFormFields, {
   type FormValues,
   type TalentRequestReferralDialogOptions,
@@ -62,7 +61,6 @@ const TalentRequestCreateTrackedUserDialog = ({
   defaultOpen = false,
 }: TalentRequestCreateTrackedUserDialogProps) => {
   const intl = useIntl();
-  const paths = useRoutes();
   const [isOpen, setOpen] = useState(defaultOpen);
   const user = getFragment(
     TalentRequestCreateTrackedUserDialog_Fragment,
@@ -146,26 +144,7 @@ const TalentRequestCreateTrackedUserDialog = ({
                 optionsQuery={optionsQuery}
                 showSelectionDecision={false}
               />
-              <Dialog.Footer>
-                <Button type="submit" disabled={fetching}>
-                  {fetching
-                    ? intl.formatMessage(commonMessages.saving)
-                    : intl.formatMessage(formMessages.saveChanges)}
-                </Button>
-                <Dialog.Close>
-                  <Button type="button" color="warning" mode="inline">
-                    {intl.formatMessage(formMessages.cancelGoBack)}
-                  </Button>
-                </Dialog.Close>
-                <Link href={paths.userView(user.id)} newTab>
-                  {intl.formatMessage({
-                    defaultMessage: "View profile",
-                    id: "KsUNz1",
-                    description:
-                      "Link text to view a user's profile in a new tab",
-                  })}
-                </Link>
-              </Dialog.Footer>
+              <ReferralDialogFooter fetching={fetching} userId={user.id} />
             </form>
           </FormProvider>
         </Dialog.Body>
